drill 4.4

+++ Chapter04/Drill.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int v1 = 0, v2 = 0;
+    double v1 = 0, v2 = 0;
     std::cout << "Please enter two integer values (enter '|' to exit):" << std::endl;
     std::cout << ">";
     while ((std::cin >> v1) && (std::cin >> v2)) {
